pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["KazuhiraBenedictMiller <andrea.amedeo.serravalle@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
dotenv-python = "^0.0.1"
pinecone-client = "^2.2.4"
langchain = "^0.0.305"
jupyter = "^1.0.0"
lab = "^7.4"
python-dotenv = "^1.0.0"
pypdf = "^3.16.2"
sentence-transformers = "^2.2.2"
torch = ">=2.0.0, !=2.0.1"
replicate = "^0.13.0"
beautifulsoup4 = "^4.12.2"
openai = "^0.28.1"
fpdf = "^1.7.2"
streamlit = "^1.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
